requirements.txt
langchain>=0.2.14
langchain-openai>=0.1.23
langchain-community>=0.2.11
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.2
tiktoken
chromadb>=0.5.5
unstructured
trafilatura
python-dotenv
pandas

